refactor(hooks): extract command-output parsers into _OutputParsingMixin
_dispatcher.py was 756 lines, mixing hook discovery / matching /
subprocess execution with ~225 lines of structured-stdout parsing
(_parse_command_output, _parse_stop_command_output). Those two methods
carry zero dispatcher state (no self.* references) — they read stdout
and mutate the passed-in result — so move them to a _OutputParsingMixin
in _output_parsing.py.

Kept as a mixin (not module functions) because tests call
dispatcher._parse_stop_command_output(...) as an instance method;
inheritance preserves that API with zero test changes. Methods moved
verbatim (semantic diff shows only the class-declaration line changed).
_dispatcher.py: 756 → 542 lines. Full suite green (2840 passed).
